+++ source/common/nats/streaming/client_pool.h
@@ -2,7 +2,7 @@
 
 #include "envoy/nats/codec.h"
 #include "envoy/nats/streaming/client.h"
-#include "envoy/tcp/conn_pool.h"
+#include "envoy/tcp/conn_pool_nats.h"
 #include "envoy/thread_local/thread_local.h"
 #include "envoy/upstream/cluster_manager.h"
